# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/companies": {
      "get": {
        "operationId": "CompaniesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List all companies",
        "tags": [
          "Companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create company",
        "tags": [
          "Companies"
        ]
      }
    },
    "/api/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get company by ID",
        "tags": [
          "Companies"
        ]
      },
      "put": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update company",
        "tags": [
          "Companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete company",
        "tags": [
          "Companies"
        ]
      }
    },
    "/api/customers": {
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/phone/{phone}": {
      "get": {
        "operationId": "CustomersController_findByPhone",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get customer by phone",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get customer by ID",
        "tags": [
          "Customers"
        ]
      },
      "put": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update customer",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete customer",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/api/inventory": {
      "get": {
        "operationId": "InventoryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List inventory records",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/low-stock": {
      "get": {
        "operationId": "InventoryController_findLowStock",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List low stock items",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/product/{productId}": {
      "get": {
        "operationId": "InventoryController_findByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inventory by product ID",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/{id}": {
      "get": {
        "operationId": "InventoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inventory record by ID",
        "tags": [
          "Inventory"
        ]
      },
      "put": {
        "operationId": "InventoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update inventory record",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/adjust": {
      "post": {
        "operationId": "InventoryController_adjustStock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Adjust inventory stock for a product",
        "tags": [
          "Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "BillBhai API",
    "description": "The BillBhai Retail Order Processing System API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
...
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "FreshKart Central",
            "description": "Company name"
          },
          "owner": {
            "type": "string",
            "example": "Ritu Malhotra",
            "description": "Owner name"
          },
          "adminName": {
            "type": "string",
            "example": "Arjun Mehta",
            "description": "Admin name"
          },
          "type": {
            "type": "string",
            "example": "Grocery Retail",
            "description": "Company type"
          },
          "email": {
            "type": "string",
            "example": "central@freshkart.in",
            "description": "Company email"
          },
          "phone": {
            "type": "string",
            "example": "+91-9870011101",
            "description": "Company phone"
          },
          "gstNo": {
            "type": "string",
            "example": "GST101",
            "description": "GST number"
          },
          "address": {
            "type": "string",
            "example": "Delhi, India",
            "description": "Company address"
          },
          "productsPlan": {
            "type": "string",
            "example": "Core POS + Inventory",
            "description": "Products plan"
          },
          "tenureMonths": {
            "type": "number",
            "example": 32,
            "description": "Tenure in months",
            "minimum": 0
          },
          "storesCount": {
            "type": "number",
            "example": 6,
            "description": "Number of stores",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "owner",
          "adminName",
          "type",
          "email",
          "phone"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "FreshKart Central",
            "description": "Company name"
          },
          "owner": {
            "type": "string",
            "example": "Ritu Malhotra",
            "description": "Owner name"
          },
          "adminName": {
            "type": "string",
            "example": "Arjun Mehta",
            "description": "Admin name"
          },
          "type": {
            "type": "string",
            "example": "Grocery Retail",
            "description": "Company type"
          },
          "email": {
            "type": "string",
            "example": "central@freshkart.in",
            "description": "Company email"
          },
          "phone": {
            "type": "string",
            "example": "+91-9870011101",
            "description": "Company phone"
          },
          "gstNo": {
            "type": "string",
            "example": "GST101",
            "description": "GST number"
          },
          "address": {
            "type": "string",
            "example": "Delhi, India",
            "description": "Company address"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "description": "Company status"
          },
          "productsPlan": {
            "type": "string",
            "example": "Core POS + Inventory",
            "description": "Products plan"
          },
          "tenureMonths": {
            "type": "number",
            "example": 32,
            "description": "Tenure in months",
            "minimum": 0
          },
          "storesCount": {
            "type": "number",
            "example": 6,
            "description": "Number of stores",
            "minimum": 0
          },
          "profit": {
            "type": "number",
            "example": 1245000,
            "description": "Profit",
            "minimum": 0
          },
          "paymentDue": {
            "type": "number",
            "example": 0,
            "description": "Payment due",
            "minimum": 0
          }
        }
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "BIZ-101",
            "description": "Company ID"
          },
          "name": {
            "type": "string",
            "example": "Rahul Sharma",
            "description": "Customer name"
          },
          "mobileNo": {
            "type": "string",
            "example": "9810001001",
            "description": "Mobile number"
          },
          "email": {
            "type": "string",
            "example": "rahul@example.com",
            "description": "Email address"
          },
          "address": {
            "type": "string",
            "example": "12, MG Road, Sector 14",
            "description": "Address"
          }
        },
        "required": [
          "companyId",
          "name",
          "mobileNo"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rahul Sharma",
            "description": "Customer name"
          },
          "mobileNo": {
            "type": "string",
            "example": "9810001002",
            "description": "Mobile number"
          },
          "email": {
            "type": "string",
            "example": "rahul@example.com",
            "description": "Email address"
          },
          "address": {
            "type": "string",
            "example": "12, MG Road, Sector 14",
            "description": "Address"
          }
        }
      },
...
      "UpdateInventoryDto": {
        "type": "object",
        "properties": {
          "stockAvailable": {
            "type": "number",
            "example": 145,
            "description": "Available stock",
            "minimum": 0
          },
          "reorderLevel": {
            "type": "number",
            "example": 20,
            "description": "Reorder level",
            "minimum": 0
          },
          "location": {
            "type": "string",
            "example": "Shelf A1",
            "description": "Storage location"
          }
        }
      },
      "AdjustStockDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "P001",
            "description": "Product ID"
          },
          "adjustment": {
            "type": "number",
            "example": 10,
            "description": "Stock adjustment (positive=restock, negative=deduct)"
          },
          "reason": {
            "type": "string",
            "example": "Restock from warehouse",
            "description": "Reason for adjustment"
          }
        },
        "required": [
          "productId",
          "adjustment"
        ]
      },
...
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}